[#329] REFACTOR: 패키지 시그니처 변경
