+++ backend/src/main/java/com/team5/secondhand/chat/notification/repository/InmemoryNotificationRepository.java
@@ -1,6 +1,6 @@
-package com.team5.secondhand.chat.noti.repository;
+package com.team5.secondhand.chat.notification.repository;
 
-import com.team5.secondhand.chat.noti.domain.SseId;
+import com.team5.secondhand.chat.notification.domain.SseKey;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
@@ -14,11 +14,11 @@ import java.util.stream.Collectors;
 @Repository
 @RequiredArgsConstructor
 public class InmemoryNotificationRepository implements NotificationRepository {
-    private final Map<SseId, SseEmitter> emitters = new ConcurrentHashMap<>();
+    private final Map<SseKey, SseEmitter> emitters = new ConcurrentHashMap<>();
     private final Map<String, Object> eventCache = new ConcurrentHashMap<>();
 
     @Override
-    public SseEmitter save(SseId id, SseEmitter sseEmitter) {
+    public SseEmitter save(SseKey id, SseEmitter sseEmitter) {
         emitters.put(id, sseEmitter);
         return sseEmitter;
     }
@@ -35,7 +35,7 @@ public class InmemoryNotificationRepository implements NotificationRepository {
     }
 
     @Override
-    public Map<SseId, SseEmitter> findAllStartById(String id) {
+    public Map<SseKey, SseEmitter> findAllStartById(String id) {
         int regIdx = id.indexOf("_");
         String prefix = id.substring(0, regIdx+1);
         return emitters.entrySet().stream()
