+++ backend/src/main/java/com/team5/secondhand/chat/notification/repository/NotificationRepository.java
@@ -1,17 +1,17 @@
-package com.team5.secondhand.chat.noti.repository;
+package com.team5.secondhand.chat.notification.repository;
 
-import com.team5.secondhand.chat.noti.domain.SseId;
+import com.team5.secondhand.chat.notification.domain.SseKey;
 import org.springframework.web.servlet.mvc.method.annotation.SseEmitter;
 
 import java.util.Map;
 import java.util.Optional;
 
 public interface NotificationRepository {
-    SseEmitter save(SseId id, SseEmitter sseEmitter);
+    SseEmitter save(SseKey id, SseEmitter sseEmitter);
 
     void deleteAllStartByWithId(String id);
 
-    Map<SseId, SseEmitter> findAllStartById(String id);
+    Map<SseKey, SseEmitter> findAllStartById(String id);
 
     Optional<SseEmitter> findStartById(String id);
 }
