+++ backend/src/main/java/com/team5/secondhand/chat/notification/service/SendChatNotificationUsecase.java
@@ -1,6 +1,6 @@
-package com.team5.secondhand.chat.noti.service;
+package com.team5.secondhand.chat.notification.service;
 
-import com.team5.secondhand.chat.noti.dto.ChatNotification;
+import com.team5.secondhand.chat.notification.dto.ChatNotification;
 
 public interface SendChatNotificationUsecase {
     void sendChatNotificationToMember(String id, ChatNotification chatNotification);
